requirements.txt
yfinance>=0.2.44
pandas>=2.0.0
numpy>=1.24.0
mplfinance>=0.12.10b0
requests>=2.31.0
python-dotenv>=1.0.0
